Reboot prevents stops devices detection
With multiple units connected to a mobly host, during the call
to get_all_instances in the AndroidDevice controller, if a device reboots,
an AdbError exception is raised which prevents the caller from receiving a
populated list of detected devices.
Updated get_instances to catch the exception to allow the population of the
device list to continue.